[read-fonts] State what can go wrong reading TrueType bytecode
The set of failures is closed by the specification rather than by any
one interpreter, and the three programs a font carries are the format's
too, so both belong beside the bytecode they describe. Nothing in this
crate executes any: they are here so that two interpreters describe the
same failure the same way.

skrifa held both, and now reads them from here. Its own copies were the
same set with the same messages, so nothing it reports changes.

`Program::table_tag` goes: only one caller, inside a `Display` that says
the same thing more directly.
